Started unit testing + refactoring the interpreter
Added NuGet package management to solution
Added reference to FakeItEasy, the best mocking lib ever!
Allow to supply the program string in the constructor as an optional
parameter or in the execute method
Moved initialization from the Execute method to the constructor
Made action list initialization into its own method.